Fix issue with removing symlinks

+++ build/postinstall-transition-1.0.0.sh
@@ -46,12 +46,16 @@ rm -rf /opt/R/arm64/share/man/man1/jags.1
 rm -rf /usr/local/bin/jags
 rm -rf /usr/local/bin/jags-uninstall
 rm -rf /usr/local/libexec/jags-terminal
-rm -rf /usr/local/include/JAGS/
+if [[ ! -h "/usr/local/include/JAGS" ]]; then
+  rm -rf /usr/local/include/JAGS/
+fi
 rm -rf /usr/local/lib/libjags.4.dylib
 rm -rf /usr/local/lib/libjags.5.dylib
 rm -rf /usr/local/lib/libjags.dylib
 rm -rf /usr/local/lib/pkgconfig/jags.pc
-rm -rf /usr/local/lib/JAGS/
+if [[ ! -h "/usr/local/lib/JAGS" ]]; then
+  rm -rf /usr/local/lib/JAGS/
+fi
 rm -rf /usr/local/lib/libjrmath.0.dylib
 rm -rf /usr/local/lib/libjrmath.la
 rm -rf /usr/local/lib/libjags.la
@@ -109,5 +113,6 @@ lns "/opt/jags/versions/jags/current-4/lib/JAGS" "/usr/local/lib/JAGS"
 lns "/opt/jags/versions/jags/current-4/lib/libjags.dylib" "/usr/local/lib/libjags.dylib"
 lns "/opt/jags/versions/jags/current-4/lib/libjags.4.dylib" "/usr/local/lib/libjags.4.dylib"
 lns "/opt/jags/versions/jags/current-4/lib/libjrmath.dylib" "/usr/local/lib/libjrmath.dylib"
+lns "/opt/jags/versions/jags/current-4/lib/libjrmath.0.dylib" "/usr/local/lib/libjrmath.0.dylib"
 
 exit $EX_OK
